requirements.txt
numpy
pandas
pillow
torch
torchvision
tqdm
scikit-learn
